refactor: 책임 분리로 인한 코드 변경
