+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,53 @@
+package roomescape.controller;
+
+import java.time.LocalTime;
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.service.ReservationTimeService;
+
+@Controller
+public class ReservationTimeController {
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    private final RowMapper<ReservationTime> actorRowMapper = (resultSet, rowNum) -> {
+        ReservationTime reservationTime = new ReservationTime(
+                resultSet.getLong("id"),
+                LocalTime.parse(resultSet.getString("start_at"))
+        );
+        return reservationTime;
+    };
+
+    @GetMapping("/times")
+    public ResponseEntity<List<ReservationTime>> read() {
+        List<ReservationTime> reservationTimes = reservationTimeService.findAllReservationTimes();
+
+        return ResponseEntity.ok().body(reservationTimes);
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTimeRequest request) {
+        ReservationTime newReservationTime = reservationTimeService.createReservationTime(request);
+
+        return ResponseEntity.ok().body(newReservationTime);
+    }
+
+    @DeleteMapping("/times/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationTimeService.deleteReservationTime(id);
+
+        return ResponseEntity.ok().build();
+    }
+}
