--- src/main/java/roomescape/controller/TimeController.java
@@ -1,71 +1,0 @@
-package roomescape.controller;
-
-import java.sql.PreparedStatement;
-import java.time.LocalTime;
-import java.util.List;
-import java.util.Objects;
-import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.domain.ReservationTime;
-
-@Controller
-public class TimeController {
-    private final JdbcTemplate jdbcTemplate;
-
-    public TimeController(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    private final RowMapper<ReservationTime> actorRowMapper = (resultSet, rowNum) -> {
-        ReservationTime reservationTime = new ReservationTime(
-                resultSet.getLong("id"),
-                LocalTime.parse(resultSet.getString("start_at"))
-        );
-        return reservationTime;
-    };
-
-    @GetMapping("/times")
-    public ResponseEntity<List<ReservationTime>> read() {
-        String sql = "select id, start_at from reservation_times";
-        List<ReservationTime> reservationTimes = jdbcTemplate.query(sql, actorRowMapper);
-
-        return ResponseEntity.ok().body(reservationTimes);
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
-        String sql = "insert into reservation_times (start_at) values (?)";
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setString(1, reservationTime.getStartAt().toString());
-            return ps;
-        }, keyHolder);
-
-        Long generatedId = Objects.requireNonNull(keyHolder.getKey()).longValue();
-        ReservationTime newReservationTime = new ReservationTime(
-                generatedId,
-                reservationTime.getStartAt()
-        );
-        return ResponseEntity.ok().body(newReservationTime);
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        String sql = "delete from reservation_times where id = ?";
-
-        jdbcTemplate.update(sql, id);
-
-        return ResponseEntity.ok().build();
-    }
-}
